requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
openpyxl==3.1.5
PyYAML==6.0.2
toml==0.10.2
requests>=2.32.0
pytest==8.3.5
testcontainers[postgres]>=4.8.0
yoyo-migrations>=8.2.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
